feat: centralize assigntodelivery flow since slack requires quick access to field names
